Make grid path finding work at runtime.
The path finding is now triggered when the "run" button is clicked, reading the roads on the map and generating the grid path.
Besides, the drop of the building blocks now changes the Iso Transform position of the blocks instead of just the transform postion, so the path finding algorithm can recognise the block's position on the map. However, the animation is currently down.
Pair programmed with Sean.